add error handling

+++ scripts/substrate-exec-rimraf.cjs
@@ -4,10 +4,16 @@ const fs = require('fs');
 
 (function() {
     const dirs = process.argv.slice(2);
+
+    if (!dirs.length) {
+        throw new Error(`rimraf-exec: Invalid args length: retrieved ${dirs.length}`)
+    }
     
     for (const dir of dirs) {
         if (fs.existsSync(dir)) {
           fs.rmSync(dir, { force: true, recursive: true });
+        } else {
+            console.warn(`rimraf-exec: ${dir} is an invalid file path and does not exist`)
         }
     }
 })();
